+++ src/Telegram.php
@@ -37,7 +37,7 @@ class Telegram
      *
      * @var string
      */
-    protected $version = '0.78.0';
+    protected $version = '0.79.0';
 
     /**
      * Telegram API key